fix: permissions cleanup

+++ Modules/Product/Http/Controllers/BrandController.php
@@ -2,7 +2,7 @@
 
 namespace Modules\Product\Http\Controllers;
 
-use App\Http\Controllers\Controller;
+use Modules\Setting\Http\Controllers\Controller;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
 use Illuminate\Foundation\Application;
@@ -19,7 +19,7 @@ class BrandController extends Controller
      */
     public function index(BrandDataTable $dataTable)
     {
-        abort_if(Gate::denies('brand.access'), 403);
+        abort_if(Gate::denies('brands.access'), 403);
         return $dataTable->render('product::brands.index');
     }
 
@@ -28,7 +28,7 @@ class BrandController extends Controller
      */
     public function create(): Factory|Application|View|\Illuminate\Contracts\Foundation\Application
     {
-        abort_if(Gate::denies('brand.create'), 403);
+        abort_if(Gate::denies('brands.create'), 403);
         return view('product::brands.create');
     }
 
@@ -37,6 +37,7 @@ class BrandController extends Controller
      */
     public function store(Request $request): RedirectResponse
     {
+        abort_if(Gate::denies('brands.create'), 403);
         $request->validate([
             'name' => 'required|string|max:255',
             'description' => 'nullable|string',
@@ -61,7 +62,7 @@ class BrandController extends Controller
      */
     public function show($id): Factory|Application|View|\Illuminate\Contracts\Foundation\Application
     {
-        abort_if(Gate::denies('brand.view'), 403);
+        abort_if(Gate::denies('brands.view'), 403);
         $brand = Brand::findOrFail($id);
         return view('product::brands.show', compact('brand'));
     }
@@ -71,7 +72,7 @@ class BrandController extends Controller
      */
     public function edit($id): Factory|Application|View|\Illuminate\Contracts\Foundation\Application
     {
-        abort_if(Gate::denies('brand.edit'), 403);
+        abort_if(Gate::denies('brands.edit'), 403);
         $brand = Brand::findOrFail($id);
         return view('product::brands.edit', compact('brand'));
     }
@@ -81,6 +82,7 @@ class BrandController extends Controller
      */
     public function update(Request $request, Brand $brand): RedirectResponse
     {
+        abort_if(Gate::denies('brands.edit'), 403);
         $request->validate([
             'name' => 'required|string|max:255',
             'description' => 'nullable|string',
@@ -101,6 +103,7 @@ class BrandController extends Controller
      */
     public function destroy(Brand $brand): RedirectResponse
     {
+        abort_if(Gate::denies('brands.delete'), 403);
         $brand->delete();
 
         // Toast a success message
